+++ src/components/common/CustomizedSnackbars.js
@@ -1,32 +1,34 @@
 import React from "react";
+
 import Snackbar from "@material-ui/core/Snackbar";
-import MuiAlert from "@material-ui/lab/Alert";
-import Slide from "@material-ui/core/Slide";
-import ClearIcon from "@material-ui/core/Icon";
+import Alert from "@material-ui/lab/Alert";
 
-function Alert(props) {
-    return <MuiAlert elevation={6} variant="filled" {...props} />;
-}
-
-function Transition(props) {
-    return <Slide {...props} direction="left" />;
-}
+// function Transition(props) {
+//     return <Slide {...props} direction="left" />;
+// }
 
 const CustomizedSnackbars = props => {
-    const { open, handleClose, type, message } = props;
+    const { open, handleClose, type ,message } = props;
 
-    return (
-        <Snackbar
-            open={open}
-            message={message}
-            onClose={handleClose}
-            TransitionComponent={Transition}
-            transitionDuration={{
-                enter: 1000,
-                exit: 1000
-            }}
-        ></Snackbar>
+    return(
+        <Snackbar open={open} autoHideDuration={6000} onClose={handleClose}>
+            <Alert onClose={handleClose} severity={type}>
+                {message}
+            </Alert>
+        </Snackbar>
     );
+    // return (
+    //     <Snackbar
+    //         open={open}
+    //         message={message}
+    //         onClose={handleClose}
+    //         TransitionComponent={Transition}
+    //         transitionDuration={{
+    //             enter: 1000,
+    //             exit: 1000
+    //         }}
+    //     ></Snackbar>
+    // );
 };
 
 export default CustomizedSnackbars;
